requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
aiofiles>=23.2.0
qdrant-client>=1.7.0
numpy>=1.24.0
PyMuPDF>=1.23.0
spacy>=3.7.0
reportlab>=4.0.0